Integrate WeatherVM ViewModel into WeatherWindow.xaml
Added a ViewModel resource (WeatherVM) to WeatherWindow.xaml and set the Grid's DataContext to this ViewModel. Updated the TextBox to bind its Text property to the Query property of the ViewModel with two-way binding. Modified the TextBlock to bind its Text property to the Query property, replacing the hardcoded city name.