Add async search functionality to WeatherVM
Added a new using directive for `WeatherApp.ViewModel.Helpers` in `WeatherVM.cs`. Introduced a new asynchronous method `MakeQuery` in the `WeatherVM` class, which calls `AccuWeatherHelper.GetCities` with the `Query` parameter. Added a new `SearchCommand` class in `SearchCommand.cs` that implements the `ICommand` interface. This class includes a constructor that takes a `WeatherVM` instance, an implementation of the `CanExecute` method that always returns `true`, and an implementation of the `Execute` method that calls the `MakeQuery` method on the `WeatherVM` instance. Added necessary using directives in `SearchCommand.cs` for namespaces such as `System`, `System.Collections.Generic`, `System.Linq`, `System.Text`, `System.Threading.Tasks`, and `System.Windows.Input`.